requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai==1.12.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0

